+++ src/Faker/Internal/Types.hs
@@ -23,6 +23,7 @@ data SourceData
   | Beer
   | BoJackHorseman
   | Book
+  | Blood
   | BossaNova
   | BreakingBad
   | Buffy